Use DOUBLE PRECISION, not REAL, for the transcript duration
The repo ports every SQLite REAL to Postgres DOUBLE PRECISION and has a
schema test that enforces it: SQLite's REAL has no exact Postgres
equivalent, so a bare REAL silently changes precision.

+++ src/mac/data/postgres/schema.sql
@@ -363,7 +363,7 @@ CREATE TABLE IF NOT EXISTS task_agent_transcripts (
     returncode INTEGER,
     started_at TEXT,
     completed_at TEXT,
-    duration_ms REAL,
+    duration_ms DOUBLE PRECISION,
     truncated INTEGER NOT NULL DEFAULT 0,
     prompt_sha256 TEXT,
     response_sha256 TEXT,
